requirements_dev.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.7

pip==18.1
bump2version==0.5.10
wheel==0.32.1
watchdog==0.9.0
flake8==3.5.0
tox==3.5.2
coverage==4.5.1
Sphinx==2.2.0

pytest==4.6.5
python-dotenv==0.13.0
pytest-dotenv==0.13.0
